Fix arm64 compilation
